requirement.txt
discord.py
llama-cpp-python
python-dotenv
